% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessNumComponents.R
\name{assessNumComponents}
\alias{assessNumComponents}
\title{Create randomly initialized models for component selection}
\usage{
assessNumComponents(
  X,
  minNumComponents = 1,
  maxNumComponents = 5,
  numRepetitions = 100,
  ctol = 1e-06,
  maxit = 2500
)
}
\arguments{
\item{X}{Input data}

\item{minNumComponents}{Minimum number of components (default 1).}

\item{maxNumComponents}{Maximum number of components (default 5).}

\item{numRepetitions}{Number of randomly initialized models to create (default 100).}

\item{ctol}{Change in SSQ needed for model to be converged (default 1e-6).}

\item{maxit}{Maximum number of iterations (default 2500).}
}
\value{
List object of the following:
\itemize{
\item metrics: metrics of every created model (number of iterations, sum of squared errors, CORCONDIA score and variance explained).
\item models: all created models.
}
}
\description{
Create randomly initialized models for component selection
}
\examples{
X = Fujita2023$data
output = assessNumComponents(X, minNumComponents=1, maxNumComponents=3, numRepetitions=10)
}
