% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corcondia.R
\name{corcondia}
\alias{corcondia}
\title{Core Consistency Diagnostic (CORCONDIA) calculation}
\usage{
corcondia(X, Fac)
}
\arguments{
\item{X}{Input data matrix}

\item{Fac}{PARAFAC model Fac object}
}
\value{
Scalar of the CORCONDIA value
}
\description{
Core Consistency Diagnostic (CORCONDIA) calculation
}
\examples{
X = Fujita2023$data
model = parafac(X, 2)
corcondia(X, model$Fac)
}
