% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correctPARAFACloadings}
\alias{correctPARAFACloadings}
\title{Transform PARAFAC loadings to an orthonormal basis.
Note: this function only works for 3-way PARAFAC models.}
\usage{
correctPARAFACloadings(dataset, model, modeToCorrect)
}
\arguments{
\item{dataset}{The input dataset used to create the model.}

\item{model}{Model output from \code{\link[=parafac]{parafac()}}.}

\item{modeToCorrect}{Correct the subject (1), feature (2) or time mode (3).}
}
\value{
Corrected loadings of the specified mode.
}
\description{
Transform PARAFAC loadings to an orthonormal basis.
Note: this function only works for 3-way PARAFAC models.
}
\examples{
library(multiway)
library(dplyr)
library(ggplot2)
library(paramGUI)
library(pracma)
set.seed(0)

# Make PARAFAC model
processedFujita = processDataCube(Fujita2023, sparsityThreshold=0.99, centerMode=1, scaleMode=2)
model = parafac(processedFujita$data, nfac=2, nstart=1, verbose=FALSE)

correctedA = correctPARAFACloadings(processedFujita, model, 1)
plot(correctedA[,1], correctedA[,2])
}
