% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exportPARAFAC}
\alias{exportPARAFAC}
\title{Export the PARAFAC model}
\usage{
exportPARAFAC(
  model,
  dataset,
  prefix = "PARAFACmodel",
  path = "./",
  saveRDS = FALSE
)
}
\arguments{
\item{model}{Model output from \code{\link[=parafac]{parafac()}}.}

\item{dataset}{A longitudinal microbiome dataset, ideally processed with \code{\link[=processDataCube]{processDataCube()}}, formatted as follows:
\describe{
\item{data}{Array object of the data cube}
\item{mode1}{Dataframe with all the subject metadata, ordered the same as the rows in the data cube.}
\item{mode2}{Taxonomic classification of the microbiota, ordered the same as the columns in the data cube.}
\item{mode3}{Dataframe with the time metadata, ordered the same as the third dimension in the array.}
}}

\item{prefix}{Prefix file name}

\item{path}{Path to a folder}

\item{saveRDS}{Save an RDS of the PARAFAC model object? (true/false, default=TRUE)}
}
\value{
Saves mode1, mode2, mode3, input data, modelled data, and data as modelled per component in .csv files.
}
\description{
Currently only supports 3-mode models.
}
\examples{
library(multiway)
library(dplyr)
library(ggplot2)
library(paramGUI)
library(pracma)
set.seed(0)

# Make PARAFAC model
processedFujita = processDataCube(Fujita2023, sparsityThreshold=0.99, centerMode=1, scaleMode=2)
model = parafac(processedFujita$data, nfac=2, nstart=1, verbose=FALSE)
\dontrun{
exportPARAFAC(model, processedFujita, prefix="Fujita")
}

}
