% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fujita2023}
\alias{Fujita2023}
\title{Fujita2023 longitudinal microbiome data}
\format{
\subsection{\code{Fujita2023}}{

A list object with three elements:
\describe{
\item{sampleMetadata}{Dataframe with all the sample metadata, ordered the same as the rows in the data cube.}
\item{taxonomy}{Taxonomic classification of the microbiota, ordered the same as the columns in the data cube.}
\item{data}{Array object of the data cube}
...
}
}
}
\source{
\url{https://doi.org/10.1186/s40168-023-01474-5}
}
\usage{
Fujita2023
}
\description{
A processed version of the Fujita2023 longitudinal microbiome dataset
The following actions have been taken to process this data:
\itemize{
\item Sparsity selection: features were kept if they were <= 99\% sparse
\item CLR transform
\item Folding into data cube
\item Centering across the subject mode
\item Scaling within the feature mode
}
}
\keyword{datasets}
