% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStabilityCheck.R
\name{modelStabilityCheck}
\alias{modelStabilityCheck}
\title{Jack-knifed PARAFAC models to determine model stability}
\usage{
modelStabilityCheck(
  X,
  sampleMetadata,
  numComponents = 1,
  numRepetitions = nrow(X)
)
}
\arguments{
\item{X}{Input data cube}

\item{sampleMetadata}{Input sample metadata}

\item{numComponents}{Number of components of the desired PARAFAC model}

\item{numRepetitions}{Number of jack-knifed samples to create}
}
\value{
List of all As, Bs, Cs of the PARAFAC models.
}
\description{
Jack-knifed PARAFAC models to determine model stability
}
\examples{
modelStability = modelStabilityCheck(Fujita2023$data, Fujita2023$sampleMetadata, numComponents=3)
}
