% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiwayCenter.R
\name{multiwayCenter}
\alias{multiwayCenter}
\title{Center a multi-way array}
\usage{
multiwayCenter(X, mode = 1)
}
\arguments{
\item{X}{Multi-way array}

\item{mode}{Mode to center across (default 1).}
}
\value{
Centered multi-way array
}
\description{
Center a multi-way array
}
\examples{
cube_cnt = multiwayCenter(Fujita2023$data)
}
