% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiwayScale.R
\name{multiwayScale}
\alias{multiwayScale}
\title{Scale a multiway array}
\usage{
multiwayScale(cube, mode = 2)
}
\arguments{
\item{cube}{Multiway array}

\item{mode}{Mode to center across: 1=subjects,2=features,3=time (default 2).}
}
\value{
Scaled multiway array
}
\description{
This function exists because \code{\link[multiway:nscale]{multiway::nscale()}} cannot handle NAs.
}
\examples{
cube_scl = multiwayCenter(Fujita2023$data)
}
