% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_fun.R
\name{parafac_fun}
\alias{parafac_fun}
\title{PARAFAC loss function calculation}
\usage{
parafac_fun(X, Fac)
}
\arguments{
\item{X}{Input data}

\item{Fac}{Fac object generated by the PARAFAC algorithm}
}
\value{
Scalar value of the loss function
}
\description{
PARAFAC loss function calculation
}
\examples{
A = array(rnorm(108*2), c(108,2))
B = array(rnorm(100*2), c(100,2))
C = array(rnorm(10*2), c(10,2))
X = reinflateTensor(A, B, C)
model = parafac(X, 2)
f = parafac_fun(X, model$Fac)
}
