% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelTCCs.R
\name{plotModelTCCs}
\alias{plotModelTCCs}
\title{Title}
\usage{
plotModelTCCs(TCC_cube)
}
\arguments{
\item{TCC_cube}{Output of [\link{assessNumComponents}]$metrics$TCC}
}
\value{
Plot
}
\description{
Title
}
\examples{
library(dplyr)
library(ggplot2)
library(ggpubr)
X = Fujita2023$data
assessment = assessNumComponents(X,
 minNumComponents = 1,
 maxNumComponents = 5,
 numRepetitions = 10,
 ctol=1e-6,
 maxit=2500)

plotModelTCCs(assessment$metrics$TCC[[2]])
}
