% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPARAFACmodel.R
\name{plotPARAFACmodel}
\alias{plotPARAFACmodel}
\title{Plot PARAFAC model}
\usage{
plotPARAFACmodel(
  model,
  colourCols = NULL,
  legendTitles = NULL,
  xLabels = NULL,
  legendColNums = NULL,
  arrangeModes = NULL,
  continuousModes = NULL,
  overallTitle = ""
)
}
\arguments{
\item{model}{Reformatted \code{\link[multiway:parafac]{multiway::parafac()}} output using \code{\link[=convertModelFormat]{convertModelFormat()}}.}

\item{colourCols}{Vector of strings stating which column names should be factorized for colours per mode.}

\item{legendTitles}{Vector of strings stating the legend title per mode.}

\item{xLabels}{Vector of strings stating the x-axis labels per mode.}

\item{legendColNums}{Vector of integers stating the desired number of columns for the legends per mode.}

\item{arrangeModes}{Vector of boolean values per mode, stating if the loadings should be arranged according to colourCols (TRUE) or not (FALSE).}

\item{continuousModes}{Vector of boolean values per mode, stating if the loadings should be plotted as a line plot (TRUE) or a bar plot (FALSE).}

\item{overallTitle}{Overall title of the plot.}
}
\value{
Plot object
}
\description{
Plot PARAFAC model
}
\examples{
library(multiway)
library(dplyr)
library(ggplot2)
set.seed(0)

# Make PARAFAC model
processedFujita = processDataCube(Fujita2023, sparsityThreshold=0.99, centerMode=1, scaleMode=2)
model = parafac(processedFujita$data, nfac=3, nstart=100, verbose=FALSE)

# Flipping the sign of some components to make the output plot equal to the paper.
model = resign(model, mode="A", absorb="C")
model = resign(model, mode="B", absorb="C")

metadataPerMode = list(processedFujita$mode1, processedFujita$mode2, processedFujita$mode3)
convertedModel = convertModelFormat(model, metadataPerMode)

# Make plot
plotPARAFACmodel(convertedModel,
  colourCols = c("", "Genus", ""),
  legendTitles = c("", "Genus", ""),
  xLabels = c("Replicate", "Feature index", "Time point"),
  legendColNums = c(0,5,0),
  arrangeModes = c(FALSE, TRUE, FALSE),
  continuousModes = c(FALSE,FALSE,TRUE),
  overallTitle = "Fujita PARAFAC model")

}
