% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDataCube.R
\name{processDataCube}
\alias{processDataCube}
\title{Process a multiway array of count data.}
\usage{
processDataCube(
  dataset,
  sparsityThreshold = 1,
  sparsityPerGroupFilter = FALSE,
  centerMode = 0,
  scaleMode = 0
)
}
\arguments{
\item{dataset}{See \link{Fujita2023}, \link{Shao2019} or \link{vanderPloeg2024}.}

\item{sparsityThreshold}{Maximum sparsity for a feature to be selected}

\item{sparsityPerGroupFilter}{Consider groups when calculating sparsity}

\item{centerMode}{Center across mode}

\item{scaleMode}{Scale within mode}
}
\value{
CLRed, centered and scaled cube
}
\description{
Process a multiway array of count data.
}
\examples{
processedCube = processDataCube(Fujita2023)
}
