% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDataToCube.R
\name{rawDataToCube}
\alias{rawDataToCube}
\title{Convert long formatted data tables (two-way) into data cubes (three-way)}
\usage{
rawDataToCube(Xlong, subjectMeta, conditionMeta, keepIndividuals)
}
\arguments{
\item{Xlong}{Tibble object}

\item{subjectMeta}{Tibble object}

\item{conditionMeta}{Tibble object}

\item{keepIndividuals}{Boolean}
}
\value{
list object with the cube and associated sample and condition metadata.
}
\description{
Convert long formatted data tables (two-way) into data cubes (three-way)
}
\examples{
output = rawDataToCube(Xlong, subjectMeta, conditionMeta, keepIndividuals=TRUE)
}
