% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reinflateBlock}
\alias{reinflateBlock}
\title{Reconstruct the data cube based on a PARAFAC model.}
\usage{
reinflateBlock(model)
}
\arguments{
\item{model}{Model output from \code{\link[=parafac]{parafac()}}.}
}
\value{
Multi-way array of the reconstructed data.
}
\description{
Note: currently only works for three-way data.
}
\examples{
model = parafac(Fujita2023$data, nfac=1, nstart=1, verbose=FALSE)
reinflatedData = reinflateBlock(model)
}
