% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reinflateTensor}
\alias{reinflateTensor}
\title{Create a tensor out of a set of matrices similar to a component model.}
\usage{
reinflateTensor(A, B, C, returnAsTensor = FALSE)
}
\arguments{
\item{A}{I x N matrix corresponding to loadings in the first mode for N components.}

\item{B}{J x N matrix corresponding to loadings in the second mode for N components.}

\item{C}{K x N matrix corresponding to loadings in the third mode for N components.}

\item{returnAsTensor}{Boolean return as rTensor S4 tensor object (default FALSE).}
}
\value{
M, an I x J x K tensor.
}
\description{
Create a tensor out of a set of matrices similar to a component model.
}
\examples{
A = rnorm(108)
B = rnorm(100)
C = rnorm(10)
M = reinflateTensor(A,B,C)
}
