% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Shao2019}
\alias{Shao2019}
\title{Shao2019 longitudinal microbiome data}
\format{
\subsection{\code{Shao2019}}{

A list object with three elements:
\describe{
\item{sampleMetadata}{Dataframe with all the sample metadata}
\item{taxonomy}{taxonomic classification}
\item{counts}{raw count data}
...
}
}
}
\source{
\url{https://doi.org/10.1038/s41586-019-1560-1}
}
\usage{
Shao2019
}
\description{
A processed version of the Shao2019 longitudinal microbiome dataset
}
\keyword{datasets}
