% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sumsqr}
\alias{sumsqr}
\title{Sum-of-squares calculation}
\usage{
sumsqr(X, na.rm = FALSE)
}
\arguments{
\item{X}{Either a list containing matrices, or a matrix of values.}

\item{na.rm}{Remove NAs from calculation (default FALSE).}
}
\value{
sum-of-squares of the object
}
\description{
Sum-of-squares calculation
}
\examples{
X = array(rnorm(108*100*10), c(108,100,10))
ssq = sumsqr(X)
}
