% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSparsity.R
\name{calculateSparsity}
\alias{calculateSparsity}
\title{Calculate sparsity across the feature mode of a multiway array}
\usage{
calculateSparsity(cube)
}
\arguments{
\item{cube}{Multiway array of counts}
}
\value{
Vector of sparsity fractions (1 x J) where J is the number of features.
}
\description{
Calculate sparsity across the feature mode of a multiway array
}
\examples{
sparsity = calculateSparsity(Fujita2023$data)
}
