% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertModelFormat}
\alias{convertModelFormat}
\title{Reformat parafac model output to plottable data.}
\usage{
convertModelFormat(model, metadataPerMode = list())
}
\arguments{
\item{model}{Model object (output of \code{\link[multiway:parafac]{multiway::parafac()}})}

\item{metadataPerMode}{List object with metadata per mode in the cube}
}
\value{
List object of the reformatted modes.
}
\description{
Reformat parafac model output to plottable data.
}
\examples{
library(multiway)
library(dplyr)
library(ggplot2)
set.seed(0)

# Make PARAFAC model
model = parafac(Fujita2023$data, nfac=3, nstart=100)

# Flipping the sign of some components to make the output plot equal to the paper.
model = resign(model, mode="A", absorb="C")
model = resign(model, mode="B", absorb="C")

# Convert model output to plottable data
subjectMetadata = Fujita2023$sampleMetadata \%>\%
filter(treat2=="WC") \%>\%
select(replicate.id) \%>\%
unique()

featureMetadata = Fujita2023$taxonomy

conditionMetadata = Fujita2023$sampleMetadata \%>\%
select(time) \%>\%
unique()

metadataPerMode = list(subjectMetadata, featureMetadata, conditionMetadata)
convertedModel = convertModelFormat(model, metadataPerMode)
}
