% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fujita2023}
\alias{Fujita2023}
\title{Fujita2023 longitudinal microbiome data}
\format{
\subsection{\code{Fujita2023}}{

A list object with three elements:
\describe{
\item{sampleMetadata}{Dataframe with all the sample metadata}
\item{taxonomy}{taxonomic classification}
\item{counts}{raw count data}
...
}
}
}
\source{
\url{https://doi.org/10.1186/s40168-023-01474-5}
}
\usage{
Fujita2023
}
\description{
A processed version of the Fujita2023 longitudinal microbiome dataset
}
\keyword{datasets}
