% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{importPARAFAC}
\alias{importPARAFAC}
\title{Import PARAFAC model}
\usage{
importPARAFAC(
  path,
  prefix,
  numComponents,
  sep = ",",
  loadRDS = TRUE,
  header = TRUE
)
}
\arguments{
\item{path}{Path to a folder}

\item{prefix}{Prefix file name}

\item{numComponents}{Number of components in the model}

\item{sep}{Separator character for the input files (default=",")}

\item{loadRDS}{Load a previously saved RDS object of the model (default=TRUE)}

\item{header}{Expect headers in the csv files (default=TRUE)}
}
\value{
List of: the PARAFAC model, the dataset used, the data as modelled
}
\description{
Import PARAFAC model
}
\examples{
library(multiway)
library(dplyr)
library(ggplot2)
library(paramGUI)
library(pracma)
set.seed(0)

# Make PARAFAC model
processedFujita = processDataCube(Fujita2023, sparsityThreshold=0.99, centerMode=1, scaleMode=2)
model = parafac(processedFujita$data, nfac=2, nstart=1, verbose=FALSE)
\dontrun{
exportPARAFAC(model, processedFujita, prefix="Fujita_")
importedModel = importPARAFAC(path="./", prefix="Fujita_", numComponents=2)
}

}
