% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStabilityCheck.R
\name{modelStabilityCheck}
\alias{modelStabilityCheck}
\title{Jack-knifed PARAFAC models to determine model stability}
\usage{
modelStabilityCheck(
  X,
  sampleMetadata,
  numComponents = 1,
  numFolds = nrow(X),
  considerGroups = FALSE,
  groupVariable = "",
  ctol = 1e-06,
  maxit = 2500
)
}
\arguments{
\item{X}{Input data cube}

\item{sampleMetadata}{Input sample metadata}

\item{numComponents}{Number of components of the desired PARAFAC model}

\item{numFolds}{Number of jack-knifed samples to create}

\item{considerGroups}{Consider subject groups in calculating sparsity (default FALSE)}

\item{groupVariable}{What column name of mode 1 should be used to consider groups? (default "")}

\item{ctol}{Change in SSQ needed for model to be converged (default 1e-6).}

\item{maxit}{Maximum number of iterations (default 2500).}
}
\value{
List of all As, Bs, Cs of the PARAFAC models.
}
\description{
Jack-knifed PARAFAC models to determine model stability
}
\examples{
processedFujita = processDataCube(Fujita2023, sparsityThreshold=0.99, centerMode=1, scaleMode=2)
modelStability = modelStabilityCheck(processedFujita$data, processedFujita$mode1, numComponents=3)
}
