% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiwayCenter.R
\name{multiwayCenter}
\alias{multiwayCenter}
\title{Center a multiway array}
\usage{
multiwayCenter(cube, mode = 1)
}
\arguments{
\item{cube}{Multiway array}

\item{mode}{Mode to center across: 1=subjects,2=features,3=time (default 1).}
}
\value{
Centered multiway array
}
\description{
This function exists because \code{\link[multiway:ncenter]{multiway::ncenter()}} cannot handle NAs.
}
\examples{
cube_cnt = multiwayCenter(Fujita2023$data)
}
