% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiwayCLR.R
\name{multiwayCLR}
\alias{multiwayCLR}
\title{Perform a centered log-ratio transform over a multiway array}
\usage{
multiwayCLR(cube)
}
\arguments{
\item{cube}{Multiway array of counts}
}
\value{
CLRed cube
}
\description{
Perform a centered log-ratio transform over a multiway array
}
\examples{
cubeCLR = multiwayCLR(Fujita2023$data)
}
