% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiwayCLR.R
\name{multiwayCLR}
\alias{multiwayCLR}
\title{Perform a centered log-ratio transform over a multi-way array}
\usage{
multiwayCLR(X, pseudocount = 1)
}
\arguments{
\item{X}{Multi-way array of counts}

\item{pseudocount}{Pseudocount value to use (default 1).}
}
\value{
CLRed cube
}
\description{
Note: Propagates NAs corresponding to missing samples.
}
\examples{
cubeCLR = multiwayCLR(Fujita2023$data)
}
