% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiwayScale.R
\name{multiwayScale}
\alias{multiwayScale}
\title{Scale a multi-way array}
\usage{
multiwayScale(X, mode = 2)
}
\arguments{
\item{X}{Multi-way array}

\item{mode}{Mode to scale within: 1=subjects,2=features,3=time (default 2).}
}
\value{
Scaled multi-way array
}
\description{
Scale a multi-way array
}
\examples{
cube_scl = multiwayCenter(Fujita2023$data)
}
