% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac4microbiome-package.R
\docType{package}
\name{parafac4microbiome-package}
\alias{parafac4microbiome}
\alias{parafac4microbiome-package}
\title{parafac4microbiome: Parallel Factor Analysis Modelling for Longitudinal Microbiome Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Creation and selection of PARAllel FACtor Analysis (PARAFAC) models of longitudinal microbiome data. You can import your own data with our import functions or use one of the example datasets to create your own PARAFAC models. Selection of the number of components is fully handled using assessModelQuality and assessModelStability giving you various metrics to select the optimal model. The selected model can then be plotted using plotPARAFACmodel which offers a 'ggplot2'-based overview of the loadings. The Parallel Factor Analysis method was originally described by Caroll & Chang (1970) \url{https://link.springer.com/article/10.1007/BF02310791} and Harshman (1970) \url{https://www.psychology.uwo.ca/faculty/harshman/wpppfac0.pdf}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://grvanderploeg.github.io/parafac4microbiome/}
}

}
\author{
\strong{Maintainer}: Geert Roelof van der Ploeg \email{g.r.ploeg@uva.nl} (\href{https://orcid.org/0009-0007-5204-3386}{ORCID}) [copyright holder]

}
\keyword{internal}
