% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelMetric.R
\name{plotModelMetric}
\alias{plotModelMetric}
\title{Plots diagnostics for selecting the number of components of a parafac model.}
\usage{
plotModelMetric(numIterations, SSE, CORCONDIA, varExp, TCC)
}
\arguments{
\item{numIterations}{Matrix of number of iterations needed per initialized model (number of models x number of components).}

\item{SSE}{Sum of squared errors of the initialized models (number of models x number of components).}

\item{CORCONDIA}{CORCONDIA scores of the initialized models (see \link{corcondia}) (number of models x number of components).}

\item{varExp}{Variation explained by the initialized models (number of models x number of components).}

\item{TCC}{Cube of Tucker Congruence Coefficients comparing the loadings between two components within a mode (mode x component x component x model).}
}
\value{
An overview plot as well as plots of the individual metrics.
}
\description{
Plots diagnostics for selecting the number of components of a parafac model.
}
\examples{
numIterations = array(round(runif(100*2)*100), c(100, 2))
SSE = array(rnorm(100*2, mean=1e4, sd=100), c(100, 2))
CORCONDIA = array(runif(100*2, min=90, max=100), c(100,2))
varExp = array(runif(100*2, min=50, max=100), c(100,2))
TCC = list(NULL, array(rnorm(3*2*2*100), c(3,2,2,100)))
plots = plotModelMetric(numIterations, SSE, CORCONDIA, varExp, TCC)
plots$overview
plots$numIterations
plots$SSE
plots$CORCONDIA
plots$varExp
}
