% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelMetric.R
\name{plotModelMetric}
\alias{plotModelMetric}
\title{Title}
\usage{
plotModelMetric(metrics, metricName = "numIterations")
}
\arguments{
\item{metrics}{Matrics object created by \code{\link[=assessNumComponents]{assessNumComponents()}}.}

\item{metricName}{Name of the metric to inspect, one of:
\itemize{
\item numIterations: the number of iterations needed to create the models (default).
\item SSE: the sum of squared error.
\item CORCONDIA: the core consistency diagnostic.
\item varExp: the variance explained (\%)
}}
}
\value{
Plot object
}
\description{
Title
}
\examples{
library(dplyr)
library(ggplot2)
X = Fujita2023$data
assessment = assessNumComponents(X, minNumComponents=1, maxNumComponents=3, numRepetitions=10)
plotModelMetric(assessment$metrics, "numIterations")
}
