% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelStability.R
\name{plotModelStability}
\alias{plotModelStability}
\title{Plot model stability check output.}
\usage{
plotModelStability(modelStabilityOutput)
}
\arguments{
\item{modelStabilityOutput}{Outout of the \code{\link[=modelStabilityCheck]{modelStabilityCheck()}} function.}
}
\value{
Plot
}
\description{
Plot model stability check output.
}
\examples{
library(dplyr)
library(ggplot2)
library(ggpubr)
modelStability = modelStabilityCheck(Fujita2023$data, Fujita2023$sampleMetadata, numComponents=3)
plotModelStability(modelStability)
}
