% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverallTCCs.R
\name{plotOverallTCCs}
\alias{plotOverallTCCs}
\title{Title}
\usage{
plotOverallTCCs(TCC_cube)
}
\arguments{
\item{TCC_cube}{Output of [\link{assessNumComponents}]$metrics$TCC}
}
\value{
Plot
}
\description{
Title
}
\examples{
library(dplyr)
library(ggplot2)
library(ggpubr)

processedFujita = processDataCube(Fujita2023,
 sparsityThreshold=0.99,
  centerMode=1,
   scaleMode=2)

assessment = assessNumComponents(processedFujita$data,
 minNumComponents = 1,
 maxNumComponents = 3,
 numRepetitions = 10,
 ctol=1e-6,
 maxit=2500)

 plotOverallTCCs(assessment$metrics$TCC[[2]])

}
