% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDataCube.R
\name{processDataCube}
\alias{processDataCube}
\title{Process a multiway array of count data.}
\usage{
processDataCube(
  dataset,
  sparsityThreshold = 1,
  considerGroups = FALSE,
  groupVariable = "",
  centerMode = 0,
  scaleMode = 0
)
}
\arguments{
\item{dataset}{See \link{Fujita2023}, \link{Shao2019} or \link{vanderPloeg2024}.}

\item{sparsityThreshold}{Maximum sparsity for a feature to be selected}

\item{considerGroups}{Consider groups when calculating sparsity}

\item{groupVariable}{Variable in mode1 data to use when considering groups}

\item{centerMode}{Center across mode (default 0, i.e. do not center)}

\item{scaleMode}{Scale within mode (default 0, i.e. do not scale)}
}
\value{
CLRed, centered and scaled cube
}
\description{
Process a multiway array of count data.
}
\examples{
processedCube = processDataCube(Fujita2023)
}
