% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reinflateBlock}
\alias{reinflateBlock}
\title{Reconstruct the data cube based on a PARAFAC model.
DEPRECATED: use \code{\link[=reinflateTensor]{reinflateTensor()}} instead.}
\usage{
reinflateBlock(Fac)
}
\arguments{
\item{Fac}{Fac object containing the components from a PARAFAC model, see \code{\link[=parafac]{parafac()}}.}
}
\value{
Multi-way array of the reconstructed data.
}
\description{
Reconstruct the data cube based on a PARAFAC model.
DEPRECATED: use \code{\link[=reinflateTensor]{reinflateTensor()}} instead.
}
\examples{
model = parafac(Fujita2023$data, nfac=1, nstart=1, verbose=FALSE)
reinflatedData = reinflateBlock(model$Fac)
}
