% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Shao2019}
\alias{Shao2019}
\title{Shao2019 longitudinal microbiome data}
\format{
\subsection{\code{Shao2019}}{

A list object with three elements:
\describe{
\item{sampleMetadata}{Dataframe with all the sample metadata, ordered the same as the rows in the data cube.}
\item{taxonomy}{Taxonomic classification of the microbiota, ordered the same as the columns in the data cube.}
\item{data}{Array object of the data cube}
...
}
}
}
\source{
\url{https://doi.org/10.1038/s41586-019-1560-1}
}
\usage{
Shao2019
}
\description{
A processed version of the Shao2019 longitudinal microbiome dataset.
The following actions have been taken to process this data:
\itemize{
\item Sparsity selection: features were kept if they were <= 90\% sparse in either birth mode group
\item CLR transform
\item Folding into data cube
\item Centering across the subject mode
\item Scaling within the feature mode
}
}
\keyword{datasets}
