% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Shao2019}
\alias{Shao2019}
\title{Shao2019 longitudinal microbiome data}
\format{
\subsection{\code{Shao2019}}{

A list object with three elements:
\describe{
\item{data}{Array object of the data cube}
\item{mode1}{Dataframe with all the subject metadata, ordered the same as the rows in the data cube.}
\item{mode2}{Taxonomic classification of the microbiota, ordered the same as the columns in the data cube.}
\item{mode3}{Dataframe with the time metadata, ordered the same as the third dimension in the array.}
...
}
}
}
\source{
\doi{10.1038/s41586-019-1560-1}
}
\usage{
Shao2019
}
\description{
The Shao2019 longitudinal microbiome dataset as a three-dimensional array,
with subjects in mode 1, microbial abundances in mode 2 and time in mode 3.
Note: only time points 4, 7, 21 and Infancy are used.
Note: all-zero microbial abundances have been removed to save disk space.
}
\keyword{datasets}
