% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vanderPloeg2024}
\alias{vanderPloeg2024}
\title{vanderPloeg2024 longitudinal microbiome data}
\format{
\subsection{\code{vanderPloeg2024}}{

A list object with three elements:
\describe{
\item{sampleMetadata}{Dataframe with all the sample metadata}
\item{taxonomy}{taxonomic classification}
\item{counts}{raw count data}
...
}
}
}
\source{
\url{https://doi.org/10.1186/s40168-023-01474-5}
}
\usage{
vanderPloeg2024
}
\description{
A processed version of the vanderPloeg2024 longitudinal microbiome dataset
}
\keyword{datasets}
