% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vanderPloeg2024}
\alias{vanderPloeg2024}
\title{vanderPloeg2024 longitudinal microbiome data}
\format{
\subsection{\code{vanderPloeg2024}}{

A list object with three elements:
\describe{
\item{data}{Array object of the data cube}
\item{mode1}{Dataframe with all the subject metadata, ordered the same as the rows in the data cube.}
\item{mode2}{Taxonomic classification of the microbiota, ordered the same as the columns in the data cube.}
\item{mode3}{Dataframe with the time metadata, ordered the same as the third dimension in the array.}
...
}
}
}
\source{
<https://doi.org/10.1101/2024.03.18.585469 >
}
\usage{
vanderPloeg2024
}
\description{
The vanderPloeg2024 longitudinal microbiome dataset as a three-dimensional array,
with subjects in mode 1, microbial abundances in mode 2, and time in mode 3.
Note: all-zero microbial abundances have been removed to save disk space.
}
\keyword{datasets}
